Move formatting logic into templates package
