+++ cli/atlas/templates/formatting_test.go
@@ -1,4 +1,4 @@
-package main
+package templates
 
 import (
 	"testing"
